Extract JWT token to it's own component
